Added wheels for aarch64 and Apple silicon chips
The default wheels are not built for aarch64, Apple M1 and M2 silicon.
https://cibuildwheel.readthedocs.io/en/stable/options/#archs

The wheels are built using the native machine. The aarch64 and universal
builds are not included in default because they cannot be tested by
native machine.

We need to set the builds specifically to support multiple
architectures.

This is the usual way and is done by other libraries too:
https://github.com/pypa/cibuildwheel/discussions/485

Also updated the version of cibuildwheel to support new architectures.