Update upstream diff-match-patch-cpp-stl to fix signed char type crash for bytestring comparisons
In #28, it was found that when using a C++ std::string type with a signed character type ('char') for the DMP library the cast to and from an unsigned integer type didn't round-trip correctly in the DMP library. This occurs when the number of unique lines in the input texts exceeds the number of positive values in the signed type. I think. We use std::string, based on char, which is signed, when comparing Python bytestrings. (Regular strings use std::wstring, which is based on wchar_t, which I think is unsigned on all platforms.)

I tried changing the C++ string type used for bytestring comparisons to wstring which fixed the cast bug and also ensured that there is space to store more than 256 unique lines in the input, as with string diffs. But performance of wstring comparisons was an order of magnitude (or more) slower. I couldn't figure out why.

The upstream change is:

```
From 673dfc9d4bcb8cd0234894f5a245e9824abf5fbc Mon Sep 17 00:00:00 2001
From: Beka Gozalishvili <beqaprogger@gmail.com>
Date: Fri, 9 Feb 2024 20:37:19 +0400
Subject: [PATCH] Fix buffer overflow in toLines function which was causing a
 fatal crash

---
 diff_match_patch.h | 3 ++-
 1 file changed, 2 insertions(+), 1 deletion(-)

diff --git a/diff_match_patch.h b/diff_match_patch.h
index 47531c9..9d184bc 100644
--- a/diff_match_patch.h
+++ b/diff_match_patch.h
@@ -682,7 +682,8 @@ class diff_match_patch {
     for (typename Diffs::iterator cur_diff = diffs.begin(); cur_diff != diffs.end(); ++cur_diff) {
       string_t text;
       for (int y = 0; y < (int)(*cur_diff).text.length(); y++) {
-        const LinePtr& lp = lineArray[static_cast<size_t>((*cur_diff).text[y])];
+          typedef typename std::make_unsigned<typename string_t::value_type>::type unsigned_value_type;
+          const LinePtr& lp = lineArray[static_cast<size_t>(static_cast<unsigned_value_type>((*cur_diff).text[y]))];
         text.append(lp.first, lp.second);
       }
       (*cur_diff).text.swap(text);
```